+++ R/pskewnorm.R
@@ -1,6 +1,6 @@
 #' Distribution function of the Skew-Normal Distribution
 #'
-#' @param x A vector of quantiles.
+#' @param q A vector of quantiles.
 #' @param par1 Location parameter of the skew-normal distribution.
 #' @param par2 Scale parameter of the skew-normal distribution.
 #' @param par3 Slant parameter of the skew-normal distribution.
@@ -18,12 +18,12 @@
 #' \emph{Cambridge University Press}.
 #'
 #' @examples
-#' pskewnorm(x = c(0.1,1.3))
+#' pskewnorm(q = c(0.1,1.3))
 #'
 #' @export
 
-pskewnorm <- function(x, par1 = 0, par2 = 1, par3 = 0) {
-  z <- (x - par1) / par2
+pskewnorm <- function(q, par1 = 0, par2 = 1, par3 = 0) {
+  z <- (q - par1) / par2
   o <- stats::pnorm(z) - 2 * sapply(z, kerTOw, a = par3)
   return(o)
 }